+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Changelog
 .. This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
+develop
+------------------
++ Fix issue with references not being read from BAM's reference binary blob if
+  ``@SQ`` header lines are not present.
+
 1.1.0
 ------------------
 + Filesizes are logged as well as space savings. Total filesizes and space